ChessProject Documentation

ChessProject.bishoplike — Method

bishoplike(board::Board)

Get the location of all the bishops and queens on the board, as a Bitboard.

ChessProject.bishops — Method

bishops(board::Board)

Get the location of all the bishops on the board, as a Bitboard.

ChessProject.cancastlekingside — Method

cancastlekingside(board::Board, color::Color)
cancastlekingside(board::Board)

Return a Bool which denotes if the player of color can castle kingside.

ChessProject.cancastlequeenside — Method

cancastlequeenside(board::Board, color::Color)
cancastlequeenside(board::Board)

Return a Bool which denotes if the player of color can castle queenside. If no color is given, it assumes the color of the current turn.

ChessProject.checkers — Method

checkers(board::Board)

Return the Bitboard of all the pieces giving check.

ChessProject.empty — Method

empty(board::Board)

Return the positions, as a Bitboard, of all the empty squares on the board.

ChessProject.enemy — Method

enemies(board::Board)

Return the positions, as a Bitboard, of all the enemies on the board.

ChessProject.friendly — Method

friendly(board::Board)

Return the positions, as a Bitboard, of all the friendly units on the board.

ChessProject.importfen — Method

importfen(fen::String)

Create a Board object from an input FEN, given as a String.

ChessProject.ischeck — Method

ischeck(board::Board)

Returns true if the king is in check, false otherwise.

ChessProject.kings — Method

kings(board::Board)

Get the location of all the kings on the board, as a Bitboard.

ChessProject.knights — Method

knights(board::Board)

Get the location of all the knights on the board, as a Bitboard.

ChessProject.monkey! — Method

monkey!(board)

The monkey AI plays random moves. It is a monkey.

ChessProject.occupied — Method

occupied(board::Board)

Return the positions, as a Bitboard, of all the occupied squares on the board.

ChessProject.pawns — Method

pawns(board::Board)

Get the location of all the pawns on the board, as a Bitboard.

ChessProject.perft — Method

perft(board::Board, depth::Int)

Run the perft routine to a given depth.

ChessProject.pinned — Method

pinned(board::Board)

Return the Bitboard of all the pieces that are pinned.

ChessProject.queens — Method

queens(board::Board)

Get the location of all the queens on the board, as a Bitboard.

ChessProject.rooklike — Method

rooklike(board::Board)

Get the location of all the rooks and queens on the board, as a Bitboard.

ChessProject.rooks — Method

rooks(board::Board)

Get the location of all the rooks on the board, as a Bitboard.

ChessProject.Bitboard — Type

Bitboard

The type used to encode the information about the 64 squares into a 64 bit number. Construct an object of type Bitboard by providing a value of type UInt64.

Example

julia> Bitboard(0x00ff00000000ff00)
Bitboard:
8
7 x x x x x x x x
6
5
4
3
2 x x x x x x x x
1
  A B C D E F G H

ChessProject.Board — Type

Board

A DataType representing the state of a chess board.

ChessProject.Color — Type

Color

DataType representing a color.

ChessProject.Magic — Type

Magic

DataType used for storing magic numbers, rook or bishop move masks, and respective shift and offset values for the computation of sliding moves.

ChessProject.Move — Type

Move

DataType used to store the information encoding a move.

ChessProject.Move — Method

Move(move_from::Integer, move_to::Integer, move_flag::Integer)

Encode a move, giving the from & to squares, alongside the promotion flag.

ChessProject.Move — Method

Move(move_from::Integer, move_to::Integer)

Encode a move, giving the from & to squares, assuming no special move flags.

ChessProject.MoveStack — Type

MoveStack

DataType for storing lists of moves.

ChessProject.Piece — Type

Piece

DataType representing a specific piece with type and color.

ChessProject.PieceType — Type

PieceType

DataType representing a type of piece.

ChessProject.Undo — Type

Undo

DataType for storing the minimal amount of information to restore a Board object to its previous position.

ChessProject.UndoStack — Type

UndoStack

DataType for storing lists of Undos.

ChessProject.BLACK — Constant

BLACK

Constant representing black.

ChessProject.WHITE — Constant

WHITE

Constant representing white.

Base.:! — Method

!(color::Color)

Find the opposite color.

Base.:& — Method

&(bb_1::Bitboard, bb_2::Bitboard)
&(bb::Bitboard, sqr::Integer)

The bitwise "and" (intersection) between two Bitboard objects. Also supports the case where one argument is a square of Integer type.

Base.:<< — Method

<<(bb::Bitboard, n::Int)

Shift the bits in the Bitboard object left by an integer n.

Base.:>> — Method

>>(bb::Bitboard, n::Int)

Shift the bits in the Bitboard object right by an integer n.

Base.:| — Method

|(bb_1::Bitboard, bb_2::Bitboard)
|(bb::Bitboard, sqr::Integer)

The bitwise "or" (union) between two Bitboard objects. Also supports the case where one argument is a square of Integer type.

Base.:~ — Method

~(bb::Bitboard)

The bitwise "not" of a Bitboard object. That is, return a Bitboard representing all the squares not in bb.

Base.:⊻ — Method

⊻(bb_1::Bitboard, bb_2::Bitboard)

The bitwise "exclusive or" between two Bitboard objects.

Base.isempty — Method

isempty(bb:Bitboard)

Determines if a given Bitboard contains any active squares

Base.isone — Method

isone(bb::Bitboard)

Returns true if the Bitboard contains only one square.

ChessProject.absearch — Method

absearch()

Naive αβ search. Under development.

ChessProject.add! — Method

add!(board::Board, piece::Piece, bb::Bitboard)
add!(board::Board, piece::Piece, sqr::Integer)

Add a piece to the board square given by an Integer sqr, or a Bitboard bb.

ChessProject.addremove! — Method

addremove!(board::Board, piece::Piece, bb::Bitboard)
addremove!(board::Board, piece::Piece, sqr::Integer)

Remove a piece from the board at the square given by an Integer sqr, or a Bitboard bb, and then add the given piece in its place.

ChessProject.apply_castle! — Method

apply_castle!(board::Board, move::Move)

Apply the given move to the board. Assumes the move has the __KING_CASTLE or __QUEEN_CASTLE flag.

ChessProject.apply_enpass! — Method

apply_enpass!(board::Board, move::Move)

Apply the given move to the board. Assumes the move has the __ENPASS flag.

ChessProject.apply_move! — Method

apply_move!(board::Board, move::Move, undo::Undo)

Apply the given move to the board, adding changes to 'undo'.

ChessProject.apply_normal! — Method

apply_normal!(board::Board, move::Move)

Apply the given move to the board. Assumes the move has either the __NORMAL_MOVE or __DOUBLE_PAWN flags.

ChessProject.apply_promo! — Method

apply_promo!(board::Board, move::Move)

Apply the given move to the board. Assumes the move has the __<PIECE>_PROMO flag, where <PIECE> is either a KNIGHT, BISHOP, ROOK, or QUEEN.

ChessProject.blockers — Method

blockers(sqr_1::Integer, sqr_2::Integer)

Retrieve all the blocking squares between sqr_1 and sqr_2, as a Bitboard.

ChessProject.chartopiece — Method

chartopiece(c::Char)

Converts a Char from the FEN to its respective Piece type.

ChessProject.color — Method

color(piece::Piece)

Return the Color type of a given piece.

ChessProject.count — Method

count(bb::Bitboard)

Count the number of filled squares in a Bitboard object.

ChessProject.evaluate — Method

evaluate(board)

Naive evaluation function to get the code development going.

ChessProject.file — Method

file(bb::Bitboard)
file(sqr::Int)

Get the Bitboard representing the file of the given square. Input is given as either an Integer type, or a Bitboard - assuming it contains only one square.

ChessProject.findpins — Method

pins(board::Board)

Retrieves all the pinned pieces on the board, as a Bitboard.

ChessProject.flag — Method

flag(move::Move)

Given a move, return any special flags, as an Integer.

ChessProject.from — Method

from(move::Move)

Given a move, retrieve the "move from" square, as an Integer.

ChessProject.gen_moves! — Method

gen_moves!(movestack::Movestack, board::Board)

Generate all the possible moves in the position.

ChessProject.gen_noisy_moves! — Method

gen_noisy_moves!(movestack::MoveStack, board::Board)

Generate all the noisy moves on a board, pushing each Move to the MoveStack.

ChessProject.gen_quiet_moves! — Method

gen_quiet_moves!(movestack::MoveStack, board::Board)

Generate all the quiet moves on a board, pushing each Move to the MoveStack.

ChessProject.isattacked — Method

isattacked(board::Board, sqr::Integer)

Is the given square attacked?

ChessProject.isdoublecheck — Method

isdoublecheck(board::Board)

Returns true if the king is in check, false otherwise.

ChessProject.ismany — Method

ismany(bb::Bitboard)

Returns true if the Bitboard contains more than one square.

ChessProject.kingAttackers — Method

kingAttackers(board::Board, sqr::Integer)

Find all the attackers of a given square, sqr.

ChessProject.kingMoves — Method

kingMoves(sqr::Integer)

Generate the potential king moves from a square given by sqr, returned as a Bitboard.

ChessProject.knightMoves — Method

knightMoves(sqr::Integer)

Generate the potential knight moves from a square given by sqr, returned as a Bitboard.

ChessProject.makepiece — Method

makepiece(piece::Piece, color::Color)

Return a piece of a given color.

ChessProject.owl! — Method

owl!(board, depth)

The owl looks ahead a move using alpha-beta pruning plus quiescence, alongside a naive evaluation. The owl is wiser than the monkey! Care should be taken in selecting the depth, as too high will take a long time.

ChessProject.pawnAdvances — Method

pawnAdvances(c::Color, sqr::Integer)

Generate the potential pawn pushes from a square given by sqr, from the point of view of the color c. Returned as a Bitboard.

ChessProject.pawnAttacks — Method

pawnAttacks(c::Color, sqr::Integer)

Generate the potential pawn attacks from a square given by sqr, from the point of view of the color c. Returned as a Bitboard.

ChessProject.piece — Method

piece(board::Board, sqr::Integer)

Get the Piece located at a square, sqr.

ChessProject.qsearch — Method

qsearch()

Quiescence search function. Under development.

ChessProject.rank — Method

rank(bb::Bitboard)
rank(sqr::Int)

Get the Bitboard representing the rank of the given square. Input is given as either an Integer type, or a Bitboard - assuming it contains only one square.

ChessProject.remove! — Method

remove!(board::Board, bb::Bitboard)
remove!(board::Board, sqr::Integer)

Remove a piece from the board at the square given by an Integer sqr, or a Bitboard bb.

ChessProject.square — Method

square(bb::Bitboard)

Returns an Int representing the first square contained within a Bitboard.

ChessProject.squareAttackers — Method

squareAttackers(board::Board, sqr::Integer)

Find all the enemy units that attack a given square, sqr.

ChessProject.switchturn! — Method

switchturn(board::Board)

Switches the Color of the side to move on a given Board object.

ChessProject.to — Method

to(move::Move)

Given a move, retrieve the "move to" square, as an Integer.

ChessProject.type — Method

type(piece::Piece)

Return the type of a given piece.

ChessProject.updatecastling! — Method

updatecastling!(board::Board, sqr_from::Integer, sqr_to::Integer)

Update the castling rights of the board, given a move is played from sqr_from to sqr_to.

ChessProject.BISHOP — Constant

BISHOP

Constant representing a bishop.

ChessProject.BISHOP_TABLE — Constant

BISHOP_TABLE

Constant containing precomputed bishop moves for all occupancy configurations.

ChessProject.EMPTY — Constant

EMPTY

A Bitboard constant representing an empty board.

ChessProject.FILE — Constant

FILE

A static vector containing all the Bitboard representations of the files of the board.

ChessProject.FILE_A — Constant

FILE_A

A Bitboard constant representing the A file of a chess board.

ChessProject.FILE_AB — Constant

FILE_AB

A Bitboard constant representing the files A and B together.

ChessProject.FILE_B — Constant

FILE_B

A Bitboard constant representing the B file of a chess board.

ChessProject.FILE_C — Constant

FILE_C

A Bitboard constant representing the C file of a chess board.

ChessProject.FILE_D — Constant

FILE_D

A Bitboard constant representing the D file of a chess board.

ChessProject.FILE_E — Constant

FILE_E

A Bitboard constant representing the E file of a chess board.

ChessProject.FILE_F — Constant

FILE_F

A Bitboard constant representing the F file of a chess board.

ChessProject.FILE_G — Constant

FILE_G

A Bitboard constant representing the G file of a chess board.

ChessProject.FILE_GH — Constant

FILE_GH

A Bitboard constant representing the files G and H together.

ChessProject.FILE_H — Constant

FILE_H

A Bitboard constant representing the H file of a chess board.

ChessProject.FULL — Constant

FULL

A Bitboard constant representing a full board.

ChessProject.KING — Constant

KING

Constant representing a King.

ChessProject.KNIGHT — Constant

KNIGHT

Constant representing a knight.

ChessProject.NONE — Constant

None

Constant representing no color.

ChessProject.PAWN — Constant

PAWN

Constant representing a pawn.

ChessProject.QUEEN — Constant

QUEEN

Constant representing a Queen.

ChessProject.RANK — Constant

RANK

A static vector containing all the Bitboard representations of the ranks of the board.

ChessProject.RANK_1 — Constant

RANK_1

A Bitboard constant representing the 1st rank (A1 -> H1) of a chess board.

ChessProject.RANK_12 — Constant

RANK_12

A Bitboard constant representing the 1st and 2nd ranks of the chess board.

ChessProject.RANK_18 — Constant

RANK_18

A Bitboard constant representing the promotion ranks of a chess board.

ChessProject.RANK_2 — Constant

RANK_2

A Bitboard constant representing the 2nd rank of a chess board.

ChessProject.RANK_27 — Constant

RANK_27

A Bitboard constant representing the starting pawn ranks of a chess board.

ChessProject.RANK_3 — Constant

RANK_3

A Bitboard constant representing the 3rd rank of a chess board.

ChessProject.RANK_4 — Constant

RANK_4

A Bitboard constant representing the 4th rank of a chess board.

ChessProject.RANK_5 — Constant

RANK_5

A Bitboard constant representing the 5th rank of a chess board.

ChessProject.RANK_6 — Constant

RANK_6

A Bitboard constant representing the 6th rank of a chess board.

ChessProject.RANK_7 — Constant

RANK_7

A Bitboard constant representing the 7th rank of a chess board.

ChessProject.RANK_78 — Constant

RANK_12

A Bitboard constant representing the 7th and 8th ranks of the chess board.

ChessProject.RANK_8 — Constant

RANK_8

A Bitboard constant representing the 8th rank of a chess board.

ChessProject.ROOK — Constant

ROOK

Constant representing a rook.

ChessProject.ROOK_TABLE — Constant

ROOK_TABLE

Constant containing precomputed rook moves for all occupancy configurations.

ChessProject.SYMBOLS — Constant

SYMBOLS

A constant holding the glyphs for each chess piece.

ChessProject.VOID — Constant

Blank

Constant representing no piece, or a blank square.